some changes

+++ implementation/data_input.cpp
@@ -79,7 +79,7 @@ dataframe data_input::getData(int col_index_of_label) {
         }
 
         // go through the parsed vector to convert from strings to numbers
-        for(int i = 0; i < vec.size(); i++){
+        for(int i = 0; i < (int)vec.size(); i++){
             //std::cout << vec.at(i) << " ";
             if(i == col_index_of_label){
                 // if we reach the index that contains the label information, place in label vector
@@ -110,7 +110,7 @@ dataframe data_input::getData(int col_index_of_label) {
             foundSizeOfInput = true;
         }else{
             // If the next line does not have the expected number of inputs,
-            if(vec.size() != expectedSizeOfInput){
+            if((int)vec.size() != expectedSizeOfInput){
                 std::cout << "Error: expected data input size of " << expectedSizeOfInput << ", got " << vec.size();
                 exit(0);
             }
